+++ packages/client/src/components/forms/LoginForm/LoginForm.tsx
@@ -3,13 +3,10 @@ import { Form, Button, Input } from "antd";
 import { useFormik } from "formik";
 import { ChangeEvent } from "react";
 import { NavLink, useNavigate } from "react-router-dom";
-import {
-    getUserInfo,
-    signin,
-    signinWithYandex,
-} from "@/services/authorization";
+import { getUserInfo, signin } from "@/services/authorization";
 import "./LoginForm.scss";
 import { LOGIN_FORM_VALIDATION_SCHEMA } from "./loginFormValidationSchema";
+import { signinWithYandex } from "@/services/oAuthYandex";
 
 export type LoginFormValuesType = {
     login: string;
