+++ packages/client/src/services/authorization.ts
@@ -1,11 +1,10 @@
 import {
-    getClientIdRequest,
     getUserDataRequest,
     signinRequest,
     signoutRequest,
     signupRequest,
 } from "@/api/Auth";
-import { signupRequestData, YandexServiceIdResponse } from "@/api/typesApi";
+import { signupRequestData } from "@/api/typesApi";
 import { LoginFormValuesType } from "@/components/forms/LoginForm/LoginForm";
 import { NavigateFunction } from "react-router-dom";
 import { MouseEventHandler } from "react";
@@ -22,26 +21,8 @@ export const signin = async (values: LoginFormValuesType) => {
     }
 };
 
-export const signinWithYandex = async () => {
-    try {
-        const response = await getClientIdRequest();
-
-        if (response.status === 200) {
-            const { service_id } = (response as YandexServiceIdResponse).data;
-
-            window.location.replace(
-                `https://oauth.yandex.ru/authorize?response_type=code&client_id=${service_id}&redirect_uri=http%3A%2F%2Flocalhost%3A3000`
-            );
-        }
-
-        return true;
-    } catch (error) {
-        return false;
-    }
-};
-
 // TODO: Добавить типизацию ответов
-export const getUserInfo = async (): Promise<any> => {
+export const getUserInfo = async () => {
     try {
         const response = await getUserDataRequest();
